Add support for travis integration
In addition to the basic .travis.yml file the changes in this branch:
- Ensure the default cucumber rake task doesn't require the browser
- Move and rename rb files to assist the rails autoloader
- Ensure that the db:seed tasks don't modify plate_purpose arrays
- Adjust some tests to be less sensitive to order of request_infos
- Avoid loading factory girl when not needed
- Removed some unused gems

More details on moving presenters:
To assist the rails autoloader, the file has been split,
and the classes have been moved under the Submission namespace
to mimic the directory layout. I've also corrected a spelling
error in the submission creator's name.
