+++ app/controllers/submissions_controller.rb
@@ -1,19 +1,20 @@
 #This file is part of SEQUENCESCAPE is distributed under the terms of GNU General Public License version 1 or later;
 #Please refer to the LICENSE and README files for information on licensing and authorship of this file.
-#Copyright (C) 2011,2012,2013,2014 Genome Research Ltd.
+#Copyright (C) 2011,2012,2013,2014,2015 Genome Research Ltd.
+
 class SubmissionsController < ApplicationController
 
   before_filter :lab_manager_login_required, :only => [:change_priority]
 
   after_filter :set_cache_disabled!, :only => [:new, :index]
-    
+
   def new
     self.expires_now
-    @presenter = SubmissionCreater.new(current_user, :study_id => params[:study_id])
+    @presenter = Submission::SubmissionCreator.new(current_user, :study_id => params[:study_id])
   end
 
   def create
-    @presenter = SubmissionCreater.new(current_user, params[:submission])
+    @presenter = Submission::SubmissionCreator.new(current_user, params[:submission])
 
     if @presenter.save
       render :partial => 'saved_order',
@@ -30,12 +31,12 @@ class SubmissionsController < ApplicationController
   end
 
   def edit
-    @presenter = SubmissionCreater.new(current_user,  :id => params[:id] )
+    @presenter = Submission::SubmissionCreator.new(current_user,  :id => params[:id] )
   end
 
   # This method will build a submission then redirect to the submission on completion
   def update
-    @presenter = SubmissionCreater.new(current_user, :id => params[:id])
+    @presenter = Submission::SubmissionCreator.new(current_user, :id => params[:id])
 
     @presenter.build_submission!
 
@@ -52,7 +53,7 @@ class SubmissionsController < ApplicationController
   def index
     # Disable cache of this page
     self.expires_now
-    
+
     @building = Submission.building.find(:all, :order => "created_at DESC", :conditions => { :user_id => current_user.id })
     @pending = Submission.pending.find(:all, :order => "created_at DESC", :conditions => { :user_id => current_user.id })
     @ready = Submission.ready.find(:all, :limit => 10, :order => "created_at DESC", :conditions => { :user_id => current_user.id })
@@ -60,7 +61,7 @@ class SubmissionsController < ApplicationController
 
   def destroy
     ActiveRecord::Base.transaction do
-      submission = SubmissionPresenter.new(current_user, :id => params[:id])
+      submission = Submission::SubmissionPresenter.new(current_user, :id => params[:id])
       if submission.destroy
         flash[:notice] = "Submission successfully deleted!"
       else
@@ -71,7 +72,7 @@ class SubmissionsController < ApplicationController
   end
 
   def show
-    @presenter = SubmissionPresenter.new(current_user, :id => params[:id])
+    @presenter = Submission::SubmissionPresenter.new(current_user, :id => params[:id])
   end
 
  def study
@@ -82,13 +83,13 @@ class SubmissionsController < ApplicationController
   ###################################################               AJAX ROUTES
   # TODO[sd9]: These AJAX routes could be re-factored
   def order_fields
-    @presenter = SubmissionCreater.new(current_user, params[:submission])
+    @presenter = Submission::SubmissionCreator.new(current_user, params[:submission])
 
     render :partial => 'order_fields', :layout => false
   end
 
   def study_assets
-    @presenter = SubmissionCreater.new(current_user, params[:submission])
+    @presenter = Submission::SubmissionCreator.new(current_user, params[:submission])
 
     render :partial => 'study_assets', :layout => false
   end
