+++ app/models/illumina_b/plate_purposes.rb
@@ -1,6 +1,7 @@
 #This file is part of SEQUENCESCAPE is distributed under the terms of GNU General Public License version 1 or later;
 #Please refer to the LICENSE and README files for information on licensing and authorship of this file.
 #Copyright (C) 2012,2013 Genome Research Ltd.
+
 module IlluminaB::PlatePurposes
   PLATE_PURPOSE_FLOWS = [
     [
@@ -70,3 +71,12 @@ module IlluminaB::PlatePurposes
   extend IlluminaHtp::PlatePurposes::PurposeHelpers
 
 end
+
+
+# We require all the plate and tube purpose files here as Rails eager loading does not play nicely with single table
+# inheritance
+
+['covaris_plate','final_plate','initial_stock_tube','mx_tube','pcr_plate','post_shear_qc_plate','stock_plate','stock_tube'].each do |type|
+  require "#{Rails.root.to_s}/app/models/illumina_b/#{type}_purpose"
+end
+
