+++ app/models/illumina_c/plate_purposes.rb
@@ -1,6 +1,7 @@
 #This file is part of SEQUENCESCAPE is distributed under the terms of GNU General Public License version 1 or later;
 #Please refer to the LICENSE and README files for information on licensing and authorship of this file.
-#Copyright (C) 2013 Genome Research Ltd.
+#Copyright (C) 2013,2015 Genome Research Ltd.
+
 module IlluminaC::PlatePurposes
   PLATE_PURPOSE_FLOWS = [
     [
@@ -82,4 +83,11 @@ module IlluminaC::PlatePurposes
     end
   end
 
+end
+
+# We require all the plate and tube purpose files here as Rails eager loading does not play nicely with single table
+# inheritance
+
+['al_libs','al_libs_tagged','lib_pcr','lib_pcr_xp','mx_tube','qc_pool','stock'].each do |type|
+  require "#{Rails.root.to_s}/app/models/illumina_c/#{type}_purpose"
 end
