+++ app/models/purpose.rb
@@ -1,6 +1,7 @@
 #This file is part of SEQUENCESCAPE is distributed under the terms of GNU General Public License version 1 or later;
 #Please refer to the LICENSE and README files for information on licensing and authorship of this file.
 #Copyright (C) 2012,2013,2015 Genome Research Ltd.
+
 class Purpose < ActiveRecord::Base
   set_table_name('plate_purposes')
