+++ app/models/search/find_illumina_b_plates_for_user.rb
@@ -1,6 +1,9 @@
 #This file is part of SEQUENCESCAPE is distributed under the terms of GNU General Public License version 1 or later;
 #Please refer to the LICENSE and README files for information on licensing and authorship of this file.
 #Copyright (C) 2012 Genome Research Ltd.
+
+require "#{Rails.root.to_s}/app/models/illumina_b/plate_purposes"
+
 class Search::FindIlluminaBPlatesForUser < Search::FindIlluminaBPlates
   def scope(criteria)
     # We find all plates that do not have transfers where they are the source.  Once a plate has been transferred (or marked
