+++ app/views/pipelines/_request_group_by_submission_inbox.html.erb
@@ -9,6 +9,7 @@
   There are no outstanding requests for this pipeline.
 <% else -%>
   <%= render :partial => 'pipeline_paginate' %>
+  <!-- _request_group_by_submission_inbox.html.erb -->
   <% form_tag({:controller => :batches, :action => :create, :id => @pipeline.id}, :id => "requests_to_batch_form") do -%>
     <% if @pipeline.item_limit -%>
       <input type="hidden" id="batch_item_limit" value="<%= @pipeline.item_limit %>" />
