+++ db/seeds/0002_plate_purposes.rb
@@ -5,7 +5,9 @@ ActiveRecord::Base.transaction do
   # And here is pulldown
   purpose_flows = Pulldown::PlatePurposes::PLATE_PURPOSE_FLOWS.clone
   purpose_flows.pop
-  purpose_flows.each do |flow|
+  purpose_flows.each do |flow_o|
+
+    flow = flow_o.clone
     # We're using a different plate purpose for each pipeline, which means we need to attach that plate purpose to the request
     # type for it.  Then in the cherrypicking they'll only be able to pick the correct type from the list.
     stock_plate_purpose = Pulldown::StockPlatePurpose.create!(
