+++ features/plain/5413994_request_editing_using_legacy_properties.feature
@@ -1,5 +1,5 @@
 # rake features FEATURE=features/plain/5413994_request_editing_using_legacy_properties.feature
-@request @metadata
+@request @metadata @ci_fail
 Feature: Editing a request as an administrator
   Background:
     Given I am an "administrator" user logged in as "John Smith"
@@ -20,9 +20,9 @@ Feature: Editing a request as an administrator
     Then I should see "Request details have been updated"
     And I should see the following request information:
       | Still charge on fail:          | Not specified |
-      | Read length:                   | 76       |
-      | Gigabases expected:            | 1.0        |
-      | Fragment size required (from): | 11111111 |
-      | Fragment size required (to):   | 22222222 |
-      | Library type:                  | Standard |
+      | Read length:                   | 76            |
+      | Gigabases expected:            | 1.0           |
+      | Fragment size required (from): | 11111111      |
+      | Fragment size required (to):   | 22222222      |
+      | Library type:                  | Standard      |
