+++ lib/tasks/cucumber.rake
@@ -1,6 +1,6 @@
 # IMPORTANT: This file is generated by cucumber-rails - edit at your own peril.
-# It is recommended to regenerate this file in the future when you upgrade to a 
-# newer version of cucumber-rails. Consider adding your own code to a new file 
+# It is recommended to regenerate this file in the future when you upgrade to a
+# newer version of cucumber-rails. Consider adding your own code to a new file
 # instead of editing this one. Cucumber will automatically load all features/**/*.rb
 # files.
 
@@ -20,6 +20,12 @@ begin
       t.profile = 'default'
     end
 
+    Cucumber::Rake::Task.new(:nojs, 'Run features without js requirements') do |t|
+      t.binary = vendored_cucumber_bin # If nil, the gem's binary is used.
+      t.fork = false # You may get faster startup if you set this to false
+      t.profile = 'nojs'
+    end
+
     Cucumber::Rake::Task.new({:wip => 'db:test:prepare'}, 'Run features that are being worked on') do |t|
       t.binary = vendored_cucumber_bin
       t.fork = true # You may get faster startup if you set this to false
@@ -36,9 +42,9 @@ begin
     task :all => [:ok, :wip]
   end
   desc 'Alias for cucumber:ok'
-  task :cucumber => 'cucumber:ok'
+  task :cucumber => 'cucumber:nojs'
 
-  task :default => :cucumber
+  task :default => 'cucumber:nojs'
 
   task :features => :cucumber do
     STDERR.puts "*** The 'features' task is deprecated. See rake -T cucumber ***"
